fix route selection
Previously, I was using MLNMapView as my base map view, but have since
switched to an NavigationMapView.

NavigationMapView supports rendering multiple routes and a selected
routes, and has a tap gesture to select a route. Things that I'd built
manually for my own MapView.

Ultimately it probably makes sense to switch to NavigationMapView's
built in handling for these things, but for now I'm going to make a
quick fix and continue to use my own route rendering and route
selection.