More useful UI after completing trip
Previously, after completing a trip, you'd just end up at the prior step, ready to start the same trip again.

But you've just completed that trip - no need to restart it.

So instead we now pop you to the "Detail" sheet of the trip arrival
screen.

And also clear the search field.