Fix intermittent hang on notched devices when searching.
I'm not exactly sure what the problem is, but it was a pretty
complicated UI Hierarchy. Namely:

HomeView is a SwiftUI view, which included...
MapView which was a UIRepresentable, wrapping...
MLNMapView which was a UIKit.UIView, which added to its subview...
TopControls which were a SwiftUI view (via a HostingViewController).

Putting TopControls as a subview of MLNMapView was maybe too much to
ask (though I don't know why it *shouldn't* work.

So now, I've pulled TopControls up a level. It's simpler and avoids the
hang.

To reproduce the hang (only on notched devices - those without touch
id):
 - tap into search bar
 - type something
 - disregard the autocomplete and hit "search"

At this point the screen would be non-interactive.